refactor: new_ui 패키지의 이름을 ui로 변경
