--- src/main/java/com/quickmemo/plugin/new_ui/button/ActionButton.java
@@ -1,7 +1,0 @@
-package com.quickmemo.plugin.new_ui.button;
-
-import com.intellij.openapi.actionSystem.AnAction;
-
-public abstract class ActionButton {
-    public abstract AnAction getCreateMemoAction();
-}
