--- src/main/java/com/quickmemo/plugin/new_ui/button/ActionListener.java
@@ -1,5 +1,0 @@
-package com.quickmemo.plugin.new_ui.button;
-
-public interface ActionListener {
-    void afterPerformed();
-}
