--- src/main/java/com/quickmemo/plugin/new_ui/memo/EmptyMemo.java
@@ -1,12 +1,0 @@
-package com.quickmemo.plugin.new_ui.memo;
-
-import com.quickmemo.plugin.memo.Memo;
-
-import java.time.LocalDateTime;
-
-public class EmptyMemo {
-    public static final Memo EMPTY_MEMO = createEmptyMemo();
-    private static Memo createEmptyMemo() {
-        return Memo.createFrom("", LocalDateTime.now());
-    }
-}
