--- src/main/java/com/quickmemo/plugin/new_ui/window/QuickMemoWindow.java
@@ -1,16 +1,0 @@
-package com.quickmemo.plugin.new_ui.window;
-
-import javax.swing.*;
-import java.awt.*;
-
-public class QuickMemoWindow {
-    private final JPanel content;
-
-    public QuickMemoWindow() {
-        this.content = new JPanel(new BorderLayout());
-    }
-
-    public JPanel getContent() {
-        return content;
-    }
-}
