+++ src/main/java/com/quickmemo/plugin/ui/button/ActionButton.java
@@ -1,0 +1,7 @@
+package com.quickmemo.plugin.ui.button;
+
+import com.intellij.openapi.actionSystem.AnAction;
+
+public abstract class ActionButton {
+    public abstract AnAction getCreateMemoAction();
+}
