+++ src/main/java/com/quickmemo/plugin/ui/toolbar/MainToolbar.java
@@ -1,0 +1,15 @@
+package com.quickmemo.plugin.ui.toolbar;
+
+import javax.swing.*;
+
+public class MainToolbar {
+    private final JPanel memoToolbar;
+
+    MainToolbar(JPanel memoToolbar) {
+        this.memoToolbar = memoToolbar;
+    }
+
+    public JPanel getMemoToolbar() {
+        return memoToolbar;
+    }
+}
