+++ src/main/resources/META-INF/plugin.xml
@@ -50,7 +50,7 @@
                     secondary="true"
                     icon="/icons/memo.svg"
                     anchor="right"
-                    factoryClass="com.quickmemo.plugin.new_ui.window.NewMemoToolWindowFactory"
+                    factoryClass="com.quickmemo.plugin.ui.window.NewMemoToolWindowFactory"
                     canCloseContents="false"/>
     </extensions>
